+++ bankfeeds/RELEASES.md
@@ -749,3 +749,13 @@ Based on:
 - [python v0.35.1] bankfeeds
 ### Releases
 - [PyPI v0.35.1] https://pypi.org/project/codat-bankfeeds/0.35.1 - bankfeeds
+
+## 2023-08-31 00:15:58
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.75.0 (2.89.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.36.0] bankfeeds
+### Releases
+- [PyPI v0.36.0] https://pypi.org/project/codat-bankfeeds/0.36.0 - bankfeeds